Move all the DRF code into separeted files
