+++ users/views.py
@@ -87,16 +87,3 @@ def panel_hide(request):
     request.user.save()
 
     return HttpResponseRedirect(reverse('index'))
-
-
-# api
-
-from users.models import User
-from rest_framework import viewsets
-from users.serializers import UserSerializer
-
-
-class UserViewSet(viewsets.ReadOnlyModelViewSet):
-    queryset = User.objects.all()
-    serializer_class = UserSerializer
-    lookup_field = 'netid'
