+++ www/urls.py
@@ -15,7 +15,6 @@ from django.views.generic import TemplateView
 from authentification import app_redirection, ulb_redirection, intranet_auth
 from django.contrib.auth.views import login, logout
 from django.contrib import admin
-from rest_framework import routers
 
 admin.autodiscover()
 
@@ -23,13 +22,6 @@ from graph.urls import json_urls as graph_json
 from views import home, node_canonic, index, p402
 import settings
 
-import users.views
-import documents.views
-
-router = routers.DefaultRouter()
-router.register(r'users', users.views.UserViewSet)
-router.register(r'documents', documents.views.DocumentViewSet)
-router.register(r'pages', documents.views.PageViewSet)
 
 # decorator whom call function_in if user is authenticated, function_out if not
 def user_logged(function_in, function_out):
@@ -41,7 +33,8 @@ def user_logged(function_in, function_out):
     return toggle
 
 
-urlpatterns = patterns("",
+urlpatterns = patterns(
+    "",
     # All JSON urls
     url(r"^json/tree/", include(graph_json)),
     url(r"^json/node/", include("polydag.urls")),
@@ -83,7 +76,7 @@ urlpatterns = patterns("",
     url(r'^help/markdown$', TemplateView.as_view(template_name='markdown.html'), name="markdown_help"),
     url(r'^help/$', TemplateView.as_view(template_name='help.html'), name="help"),
 
-    url(r'^api/', include(router.urls)),
+    url(r'^api/', include("www.rest_urls")),
 )
 
 handler400 = 'www.error.error400'
